Move RequestsSender to its own package
