+++ app/src/main/java/com/rebalance/backend/service/BackendService.kt
@@ -5,7 +5,7 @@ import android.os.StrictMode
 import android.util.Base64
 import com.google.gson.Gson
 import com.rebalance.PreferencesData
-import com.rebalance.backend.api.*
+import com.rebalance.backend.api.request.*
 import com.rebalance.backend.entities.*
 import com.rebalance.backend.exceptions.ServerException
 import kotlinx.parcelize.Parcelize
